Added more features
Preparing a TemporaryProtection for PlayerProtection, it'll be installed
on BlockController.
It is supposed to hold  player blocks until player put more blocks near
the one.

Added ban system, some commands, bug fixes.
Nice feature: gamemode players and moderators can't drop "safe" items,
that is: players that can modify the course of the game will be isolated